Add Score Save
Simple binary "best score" saving

+++ Hot Ball/Assets/Code/Score/ScoreCount.cs
@@ -3,7 +3,7 @@
 public class ScoreCount
 {
     private int _score;
-    public float ScoreGame { get => _score; }
+    public int ScoreGame { get => _score; }
 
     public void FixedUpdate() =>
         _score = (int)(Time.time * Constants.NiceScoreCount);
